Update 1.1.0
Mostly added new configurations and a fix for the suspicion meter of the T-Rex.

- Suspicion:
	- Fixed the T-Rex's suspicion meter. The T-Rex already had a suspicion meter but it would decrease it's suspicion level too early and fast.
-Configuration:
	- New configurations have been added for the T-Rex's behavior:
		- Vision Range length
		- Vision Range Width
		- Max Suspicion Level
		- Suspicion Increment
		- Suspicion Decrement
		- Suspicion Decrease Time

These changes should make the T-Rex less "blind" as the suspicion is no longer decreased too early after just increasing. The configurations will also help as the T-Rex's vision can now be adjusted for those who want it to be more "aggresive" or dangerous.